Set stdin to null when starting a wasmcloud host with wash-lib
By default `tokio::process::Command` inherits stdin from the running
process, which means that if you're trying to start a wasmcloud host
with wash-lib it will capture stdin and make that particular terminal
unusable until you kill the wasmcloud process.

This fixes the issue by setting stdin to null when starting the process.

Signed-off-by: Dan Norris <protochron@users.noreply.github.com>